rename: HomeViewModel to VoiceViewModel
